refactor(frontend): use* 関数の格納場所を正式名称（composables）に変更
