+++ packages/frontend/src/composables/use-chart-tooltip.ts
