+++ packages/frontend/src/composables/use-leave-guard.ts
