+++ packages/frontend/src/composables/use-note-capture.ts
