+++ packages/frontend/src/composables/use-scroll-position-keeper.ts
